+++ sdks/typescript/api/resources/addresses/client/requests/index.ts
@@ -1,3 +1,2 @@
 export { type GetOrganizationAddressesRequest } from "./GetOrganizationAddressesRequest.js";
 export { type GetProfileAddressesRequest } from "./GetProfileAddressesRequest.js";
-export { type CreateAddressRequest } from "./CreateAddressRequest.js";