+++ sdks/typescript/api/resources/assets/client/requests/index.ts
@@ -1,2 +1,1 @@
-export { type GetAssetsByAddressRequest } from "./GetAssetsByAddressRequest.js";
-export { type GetAssetsByWalletRequest } from "./GetAssetsByWalletRequest.js";
+export { type GetAssetsRequest } from "./GetAssetsRequest.js";