+++ sdks/typescript/api/resources/wallets/client/requests/CreateWalletRequest.ts
@@ -17,7 +17,9 @@ export interface CreateWalletRequest {
 export namespace CreateWalletRequest {
     /** Blockchain network */
     export const Network = {
+        Btc: "BTC",
         TBtc: "tBTC",
+        Eth: "ETH",
         TEth: "tETH",
     } as const;
     export type Network = (typeof Network)[keyof typeof Network];